add patch 007: remove 480 character limit for destination file path length
the patch adds a different limit instead: when option -s is used for
splitting games, there is now a maximum number of games that can be
split. the limit is a compile time constant. the default limit is
999999 games, but it can be changed by defining the preprocessor
macro MAXIMUM_NUMBER_OF_GAMES. the value must fit into an int type.

before the patch there was a theoretical limit of 999999999999999
(15 digits) games that could be split before the sprintf() string
formatting would exceed the string buffer limit and cause a buffer
overrun. in practice this limit was unreachable because the counter
variable of type int would overflow to a negative number before reaching
that number.