ISgfcSinglePropertyValue is now virtual inherited everywhere
because C++ does not have a dedicated interface feature, the current
implementation is forced to employ multiple inheritance. the interface
class ISgfcSinglePropertyValue is inherited multiple times in all
typed property value classes:

- once via the SgfcSinglePropertyValue base class
- once via the typed interface class (e.g. ISgfcColorPropertyValue)

in this commit we make sure that all sub-classes of
ISgfcSinglePropertyValue use virtual inheritance. this makes sure that
objects contain only one shared ISgfcSinglePropertyValue instance.

without virtual inheritance a class such as SgfcColorPropertyValue
cannot be instantiated because the compiler sees it as abstract, because
SgfcColorPropertyValue does not implement all pure virtual methods that
it inherits via the ISgfcColorPropertyValue base class.