add Unknown to SgfcGameType enum; remove numeric values from all enum members
the idea for the numeric values was to make it easier to map the raw
string property values in the SGF content to the corresponding enum
members, and vice versa.

with the addition of the new enum member Unknown this scheme becomes
error prone: Unknown does not have a numeric representation in the SGF
content, so every piece of code that uses the numeric value of enum
members would now have to add a special check to guard against Unknown.
the risk of something going wrong may be manageable within the library,
but SgfcGameType is part of the library's public API, so library clients
might start to rely on the numeric values, too.

to prevent this from happening, the numeric values are removed.