SgfcDocumentWriter now populates the SgfcBackendDataWrapper object with SGF content
this is the final piece in the puzzle that joins the stuff implemented
in the previous commits: it uses the content generated by
SgfcDocumentEncoder to populate a newly allocated SgfcBackendDataWrapper
object. the SgfcBackendController picks up that SGF content and injects
it into SGFC for parsing and saving to an .sgf file.