SgfcBoardSizeProperty::GetBoardSize can now return SgfcConstants::BoardSizeInvalid
with this change it is now possible to distinguish between the case
where the board size value is actually missing
(SgfcConstants::BoardSizeNone),
and the case where a board size value is present but it is invalid
(SgfcConstants::BoardSizeInvalid). before the two cases were lumped
together under the return value SgfcConstants::BoardSizeNone.