handle multi-threaded access to SGFC in SgfcBackendController
this protection is necessary because SGFC has many global and static
variables. because all access to SGFC goes through SgfcBackendController
the addition of multi-threading support is a relatively simple manner:
just add a static mutex and lock it whenever SgfcBackendController wants
to invoke any of SGFC's functions.

the library itself has no global variables. it has a number of static
variables, but these are all constants that are initialized when the
runtime starts up. SGFC is the only thing that needs to be protected
against access from multiple threads.