Doxygen now genrates documentation for exported enum class types (#13)
the preprocessor macro SGFCPLUSPLUS_EXPORT is defined in file
SgfcPlusPlusExport.h, which is generated by CMake. Doxygen does not know
where to look for SgfcPlusPlusExport.h, so it does not know the value
of SGFCPLUSPLUS_EXPORT. for regular classes this does not seem to be a
problem, but for C++11 enum classes Doxygen somehow can't deal with the
unresolved macro. the result is that Doxygen prints this warning
  documented symbol ... was not declared or defined
and does not generate documentation for the enumeration.

one way how to solve this is to add the search path where to look for
SgfcPlusPlusExport.h in the INCLUDE_PATH tag. the problem with this is
that there is no well known path because whoever builds the library is
free to set up the build environment after his or her own taste. also
it might be desirable to generate the docs without having previously
built the software, which means that SgfcPlusPlusExport.h might not
exist at all.

a better solution therefore is to configure Doxygen to have its own
definition of SGFCPLUSPLUS_EXPORT. since the macro value is not
relevant for the generated documentation the macro can be simply defined
without a value.