overhaul of how property value objects are created
an early design decision was that if during SGF parsing a raw string
value could not be converted to its proper type the parsing process
should still create a property value object of the appropriate type.

example: if a property value was expected to be a Number value, but
the raw string value could not be converted to an SgfcNumber, the
resulting property value object should still be an instance of
ISgfcNumberPropertyValue.

this decision turned out to be flawed because it led to awkward
behaviour in the property value classes. for instance it is natural
that an ISgfcNumberPropertyValue object can provide an SgfcNumber
value, but it was forbidden to do so if conversion from the raw string
value failed - the GetNumberValue() method was forced to throw an
exception.

this madness propagated all the way to the Go specific property value
classes, and from there even to the basic data classes such as
ISgfcGoStone, which had to support a "null" location in case the
stone's location on the board could not be parsed for some reason.

this commit puts a stop to all this.
- property value classes can no longer be constructed without an
  actual typed value. if SgfcPropertyDecoder somehow fails to determine
  a typed value it must instead construct an SgfcSinglePropertyValue
  object. that object's GetValueType() still reports the expected
  value type, but subclasses no longer have to deal with the awkward
  "I have no value" situation. for instance, GetNumberValue(),
  GetColorValue() etc. no longer throw std::logic_error.
- SgfcSinglePropertyValue is now no longer an abstract class - this no
  longer works because it must be possible to create instances of
  SgfcSinglePropertyValue. SgfcSinglePropertyValue now has a member
  variable that stores the SgfcPropertyValueType. if a subclass is
  instantiated it must pass the SgfcSinglePropertyValue that is
  appropriate for its class.
- the dummy class SgfcUnknownPropertyValue has been removed because it
  is no longer useful, because SgfcSinglePropertyValue is no longer
  abstract and can directly store SgfcPropertyValueType::Unknown.
- ISgfcGoStone and ISgfcGoMove can no longer live without an
  ISgfcGoPoint; this makes the the Go property value interfaces much
  saner
- Go property value classes now have simple constructors that
  require their corresponding value object (e.g.
  ISgfcGoPointPropertyValue now requires an ISgfcGoPoint). there is
  now slightly more work for SgfcPropertyDecoder and
  SgfcPropertyValueFactory, but the property value classes have become
  simpler to understand
- the documentation of ISgfcSinglePropertyValue has been overhauled
  to clarify the two creation scenarios (create by library client,
  create by libsgfc++ during parsing)
- implementation of ISgfcBoardSizeProperty and ISgfcGameTypeProperty
  has become slightly simpler because it is now no longer possible to
  have ISgfcNumberPropertyValue objects that don't have a Number value