revert example program back to -E1 (#31)
it is unlikely that modern-day SGF-writing application are actually
following the SGF standard's specification how encoding should be done.
the problem is escaping in conjunction with multi-byte encodings used
in Asian countries.

so although -E1 is not strictly conforming to the SGF standard, it is
the better option than -E2.