escape processing is now optional and default-disabled in SgfcProperyDecoder (#31)
the reason for this change is that SGFC V2.00 now removes escape
characters for us, so we don't need to do the unescaping ourselves
any more.

the code is too valuable to discard, so it is left in but made optional,
just in case it will be needed again at some time in the future.