re-implement tree traversal without recursion to avoid stack overflow
when large .sgf files with trees deeper than 1000 are processed, a
recursion-based implementation is likely to encounter a stack overflow.

use the example program to parse the test file in this commit to prove
that libsgfc++ can process game trees that are many thousand nodes
deep.

  ./libsgfcplusplus-example docread deep-game-tree.sgf

this generates several MB of rather useless output. what is interesting
about the output is that the program does not crash, and that it prints
the correct number of nodes: 1000, 2000 and 5000.